pyproject.toml
[tool.poetry]
name = "insight"
version = "0.1.0"
description = "Web app. Allows a user to upload data from Google Sheets and make sense of it."
authors = ["Sam Kitonyi <sam.kitonyi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
dash = "^1.18.1"
pandas = "^1.2.0"
google-api-python-client = "^1.12.8"
python-dotenv = "^0.15.0"
gunicorn = "^20.0.4"
dash-extensions = "^0.0.44"
openpyxl = "^3.0.6"
Flask-SQLAlchemy = "^2.4.4"
Flask-Login = "^0.5.0"
Flask-Migrate = "^2.5.3"
psycopg2-binary = "^2.8.6"
dash-bootstrap-components = "^0.11.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
pylint = "^2.6.0"
rope = "^0.18.0"
ipykernel = "^5.4.3"
mypy = "^0.790"
nbformat = "^5.1.2"
Flask-Testing = "^0.8.1"
requests = "^2.25.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
